Rename single prescription hook
