+++ client/packages/invoices/src/Prescriptions/DetailView/PrescriptionLineEdit/PrescriptionLineEditForm.tsx
@@ -19,7 +19,7 @@ import {
   StockItemSearchInput,
   ItemRowFragment,
 } from '@openmsupply-client/system';
-import { usePrescriptionSingle } from '../../api';
+import { usePrescription } from '../../api';
 import { DraftItem } from '../../..';
 import { PackSizeController } from '../../../StockOut';
 import {
@@ -74,7 +74,7 @@ export const PrescriptionLineEditForm: React.FC<
   const [issueQuantity, setIssueQuantity] = useState(0);
   const { format } = useFormatNumber();
   // const { items } = usePrescription.line.rows();
-  const { rows: items } = usePrescriptionSingle();
+  const { rows: items } = usePrescription();
 
   const onChangePackSize = (newPackSize: number) => {
     const packSize = newPackSize === -1 ? 1 : newPackSize;
